Move block by regex to Extra from Groups.
