+++ app/src/main/java/me/lucky/silence/fragment/ExtraFragment.kt
@@ -8,6 +8,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.core.widget.doAfterTextChanged
 import androidx.fragment.app.Fragment
 
 import me.lucky.silence.Preferences
@@ -39,6 +40,7 @@ class ExtraFragment : Fragment() {
             plusNumbers.isChecked = prefs.isBlockPlusNumbers
             stir.isEnabled = Build.VERSION.SDK_INT >= Build.VERSION_CODES.R
             stir.isChecked = prefs.isStirChecked
+            txtLayoutRegex.editText?.setText(prefs.regexPattern)
         }
     }
 
@@ -59,6 +61,32 @@ class ExtraFragment : Fragment() {
         stir.setOnCheckedChangeListener { _, isChecked ->
             prefs.isStirChecked = isChecked
         }
+
+        txtLayoutRegex.editText?.doAfterTextChanged {
+            val txt = it?.toString()
+            if(isBlockRegexValid(txt)) {
+                prefs.regexPattern = txt
+                txtLayoutRegex.error = ""
+            }
+            else {
+                txtLayoutRegex.error = "Invalid regex!"
+            }
+
+            return@doAfterTextChanged
+        }
+    }
+
+    private fun isBlockRegexValid(pattern: String?): Boolean {
+        if (pattern.isNullOrBlank())  return true
+
+        try {
+            pattern.toRegex()
+        }
+        catch(ex: java.util.regex.PatternSyntaxException) {
+            //Toast.makeText(this.ctx, ex.message, Toast.LENGTH_LONG).show()
+            return false
+        }
+        return true
     }
 
     private val registerForContactsPermissions =
@@ -66,4 +94,6 @@ class ExtraFragment : Fragment() {
 
     private fun requestContactsPermissions() =
         registerForContactsPermissions.launch(Manifest.permission.READ_CONTACTS)
+
+
 }
