+++ app/src/main/java/me/lucky/silence/fragment/GroupsFragment.kt
@@ -40,7 +40,6 @@ class GroupsFragment : Fragment() {
             tollFree.isChecked = value.and(Group.TOLL_FREE.value) != 0
             mobile.isChecked = value.and(Group.MOBILE.value) != 0
             localMobile.isChecked = value.and(Group.LOCAL_MOBILE.value) != 0
-            txtLayoutRegex.editText?.setText(prefs.regexPattern)
         }
     }
 
@@ -60,31 +59,5 @@ class GroupsFragment : Fragment() {
         localMobile.setOnCheckedChangeListener { _, isChecked ->
             prefs.groups = Utils.setFlag(prefs.groups, Group.LOCAL_MOBILE.value, isChecked)
         }
-
-        txtLayoutRegex.editText?.doAfterTextChanged {
-            val txt = it?.toString()
-            if(isBlockRegexValid(txt)) {
-                prefs.regexPattern = txt
-                txtLayoutRegex.error = ""
-            }
-            else {
-                txtLayoutRegex.error = "Invalid regex!"
-            }
-
-            return@doAfterTextChanged
-        }
-    }
-
-    private fun isBlockRegexValid(pattern: String?): Boolean {
-        if (pattern.isNullOrBlank())  return true
-
-        try {
-            pattern.toRegex()
-        }
-        catch(ex: java.util.regex.PatternSyntaxException) {
-            //Toast.makeText(this.ctx, ex.message, Toast.LENGTH_LONG).show()
-            return false
-        }
-        return true
     }
 }
