+++ app/src/main/res/layout/fragment_extra.xml
@@ -103,7 +103,37 @@
                 android:layout_height="wrap_content"
                 android:text="@string/extra_stir_description"
                 android:textAppearance="?attr/textAppearanceBodySmall" />
+            <Space
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginVertical="5dp" />
 
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="34dp"
+                android:text="@string/groups_match_regex"
+                android:textAppearance="?attr/textAppearanceBodyLarge" />
+
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/groups_match_regex_description"
+                android:textAppearance="?attr/textAppearanceBodySmall" />
+            <com.google.android.material.textfield.TextInputLayout
+                android:id="@+id/txtLayoutRegex"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:hint="@string/groups_match_regex_hint">
+
+                <com.google.android.material.textfield.TextInputEditText
+                    android:id="@+id/txtInputRegex"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:inputType="text"
+                    android:singleLine="true"
+                    android:textAppearance="?attr/textAppearanceCaption" />
+            </com.google.android.material.textfield.TextInputLayout>
         </LinearLayout>
     </ScrollView>
 </FrameLayout>
